files-ms/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-multipart
kafka-python